* Move header file 'GPUInfo.h' to 'common'
