+++ runtime/common/GPUInfo.h
